bug fix where 'php artisan migrate:reset' would error.

+++ migrations/default/2023_03_24_125122_add_id_to_related.php
@@ -27,8 +27,10 @@ return new class extends Migration {
 
     public function down(): void
     {
-        Schema::table(config('twill.related_table', 'twill_related'), function (Blueprint $table) {
-            $table->dropColumn('id');
-        });
+      if (Schema::hasColumn(config('twill.related_table', 'twill_related'), 'id')) {
+            Schema::table(config('twill.related_table', 'twill_related'), function (Blueprint $table) {
+                $table->dropColumn('id');
+            });
+        }
     }
 };
